% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCCAQv2.R
\name{bccaqv2.netcdf.wrapper}
\alias{bccaqv2.netcdf.wrapper}
\title{Wrapper function for the entire BCCAQv2 downscaling method}
\usage{
bccaqv2.netcdf.wrapper(gcm.file, obs.file, out.file, varname = "tasmax")
}
\arguments{
\item{gcm.file}{Filename of GCM simulations in NetCDF format}

\item{obs.file}{Filename of high-res gridded historical observations}

\item{out.file}{The file to create (or overwrite) with the final BCCAQv2 NetCDF output}

\item{varname}{Name of the NetCDF variable to downscale (e.g. 'tasmax')}
}
\description{
This function does everything
}

