% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimDown.R
\docType{package}
\name{ClimDown}
\alias{ClimDown}
\alias{ClimDown-package}
\title{ClimDown: PCIC's daily Climate Downscaling library}
\description{
This package includes PCIC's routines and techniques for
downscaling coarse scale Global Climate Models (GCMs) to fine scale
spatial resolution.
}
\details{
At present, the package only exports high-level wrapper function
that perform each of three downscaling steps: BCCI, BCCA, and
QPQM. Each wrapper simply takes four arguments: GCM file, gridded
observation file, output file, and variable name.

The package also provides three wrapper scripts that allow the user
to run each step from the command line with Rscript.
}
\references{
Alex J. Cannon, Stephen R. Sobie, and Trevor Q. Murdock, 2015: Bias Correction of GCM Precipitation by Quantile Mapping: How Well Do Methods Preserve Changes in Quantiles and Extremes?. J. Climate, 28, 6938–6959.
}
\keyword{climate}
\keyword{downscaling}

