% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPQM.R
\name{qpqm.netcdf.wrapper}
\alias{qpqm.netcdf.wrapper}
\title{High-level wrapper for Quantile perturbation quantile mapping (QPQM)}
\usage{
qpqm.netcdf.wrapper(obs.file, gcm.file, out.file, varname = "tasmax")
}
\arguments{
\item{obs.file}{Filename of high-res gridded historical observations}

\item{gcm.file}{Filename of GCM simulations interpolated to the obs.file grid}

\item{out.file}{Filename to create (or overwrite) with the bias corrected outputs}

\item{varname}{Name of the NetCDF variable to downscale (e.g. 'tasmax')}
}
\description{
This function performs the QPQM algorithm on a
cell-by-cell basis for each cell in the spatial domain of the
inputted high-res gridded observations. It uses the gridded
observations plus the GCM-based output of BCCI as input to the
algorithm and then performs a quantile perturbation/quantile
mapping bias correction. The output is written out to out.file.
}

