% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimDown.R
\docType{package}
\name{ClimDown}
\alias{ClimDown}
\alias{ClimDown-package}
\title{ClimDown: PCIC's daily Climate Downscaling library}
\description{
This package includes PCIC's routines and techniques for
downscaling coarse scale Global Climate Models (GCMs) to fine scale
spatial resolution.
}
\details{
At present, the package only exports high-level wrapper function
that perform each of three downscaling steps: CI, CA, and QDM, as
well as one wrapper that runs the entire pipeline: BCCAQ. In
general, each wrapper simply takes four arguments: GCM file,
gridded observation file, output file, and variable name. However,
see the specific function documentation for specifics.

The package also provides five wrapper scripts that allow the user
to run each step from the command line (plus the whole pipeline)
with Rscript.
}
\references{
Werner, A. T., & Cannon, A. J. (2016). Hydrologic extremes  - an intercomparison of multiple gridded statistical downscaling methods. Hydrology and Earth System Sciences, 20(4), 1483-1508. doi: 10.5194/hess-20-1483-2016
}
\keyword{climate}
\keyword{downscaling}

