% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDM.R
\name{qdm.netcdf.wrapper}
\alias{qdm.netcdf.wrapper}
\title{High-level NetCDF wrapper for Quantile Delta Mapping (QDM)}
\usage{
qdm.netcdf.wrapper(qpqm.file, obs.file, analogues, out.file,
  varname = "tasmax")
}
\arguments{
\item{qpqm.file}{The output file from the QPQM script}

\item{obs.file}{Filename of high-res gridded historical observations}

\item{analogues}{Temporal analogues... describe this more}

\item{out.file}{The file to create (or overwrite) with the final NetCDF output}

\item{varname}{Name of the NetCDF variable to downscale (e.g. 'tasmax')}
}
\description{
All files (save for the analogues_file) should have the same spatial domain.
}

